\name{aaacgj2022}
\docType{data}
\alias{aaacgj2022}
\title{Dataset of Adekoya, Akiseye, Antonakakis, Chatziantoniou, Gabauer and Oliyide (2022)}
\description{
For detailed information see: Adekoya, O. B., Akinseye, A., Antonakakis, N., Chatziantoniou, I., Gabauer, D., & Oliyide, J. A. (2022). Crude oil and Islamic sectoral stocks: Asymmetric connectedness and investment strategies. Available at SSRN.
}
\usage{data(aaacgj2022)}
\format{zoo data.frame}
\keyword{datasets}
