\name{aaacgo2022}
\docType{data}
\alias{aaacgo2022}
\title{Dataset of Adekoya, Akinseye, Antonakakis, Chatziantoniou, Gabauer and Oliyide (2022)}
\description{
For detailed information see: Adekoya, O. B., Akinseye, A., Antonakakis, N., Chatziantoniou, I., Gabauer, D., and Oliyide, J. A. (2021). Crude oil and Islamic sectoral stocks: Asymmetric connectedness and investment strategies. Available at SSRN.
}
\usage{data(aaacgo2022)}
\format{zoo data.frame}
\keyword{datasets}
