\name{acfgg2020}
\docType{data}
\alias{acfgg2020}
\title{Dataset of Antonakakis, Cunado, Filis, Gabauer and Gracia (2020)}
\description{
For detailed information see: Antonakakis, N., Cunado, J., Filis, G., Gabauer, D., & de Gracia, F. P. (2020). Oil and asset classes implied volatilities: Investment strategies and hedging effectiveness. Energy Economics, 91, 104762.
}
\usage{data(acfgg2020)}
\format{zoo data.frame}
\keyword{datasets}
