\name{acg2020}
\docType{data}
\alias{acg2020}
\title{Dataset of Antonakakis, Chatziantoniou and Gabauer (2020)}
\description{
For detailed information see: Antonakakis, N., Chatziantoniou, I., & Gabauer, D. (2020). Refined measures of dynamic connectedness based on time-varying parameter vector autoregressions. Journal of Risk and Financial Management, 13(4), 84.
}
\usage{data(acg2020)}
\format{zoo data.frame}
\keyword{datasets}
