% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorBayes.R
\name{BayesPrior}
\alias{BayesPrior}
\title{Bayes Prior}
\usage{
BayesPrior(x, nlag)
}
\arguments{
\item{x}{zoo data matrix}

\item{nlag}{Lag length}
}
\value{
Get VAR coefficients and variance-covariance matrix
}
\description{
Get Bayes Prior
}
\examples{
#data(dy2012)
#prior = BayesPrior(dy2012, nlag=1)
}
\references{
Primiceri, G. E. (2005). Time varying structural vector autoregressions and monetary policy. The Review of Economic Studies, 72(3), 821-852.
}
\author{
David Gabauer
}
