\name{bcg2022}
\docType{data}
\alias{bcg2022}
\title{Dataset of Broadstock, Chatziantoniou and Gabauer (2022)}
\description{
For detailed information see: Broadstock, D., Broadstock, D. C., Chatziantoniou, I., & Gabauer, D. (2022). Minimum connectedness portfolios and the market for green bonds: Advocating socially responsible investment (SRI) activity. In Applications in Energy Finance (pp. 217-253). Palgrave Macmillan, Cham.
}
\usage{data(bcg2022)}
\format{zoo data.frame}
\keyword{datasets}
