\name{bgu2021}
\docType{data}
\alias{bgu2021}
\title{Dataset of Balcilar, Gabauer and Umar (2021)}
\description{
For detailed information see: Balcilar, M., Gabauer, D., & Umar, Z. (2021). Crude Oil futures contracts and commodity markets: New evidence from a TVP-VAR extended joint connectedness approach. Resources Policy, 73, 102219.
}
\usage{data(bgu2021)}
\format{zoo data.frame}
\keyword{datasets}
