\name{cg2021}
\docType{data}
\alias{cg2021}
\title{Dataset of Chatziantoniou and Gabauer (2021)}
\description{
For detailed information see: Chatziantoniou, I., & Gabauer, D. (2021). EMU risk-synchronisation and financial fragility through the prism of dynamic connectedness. The Quarterly Review of Economics and Finance, 79, 1-14.
}
\usage{data(cg2021)}
\format{zoo data.frame}
\keyword{datasets}
