\name{cgg2022}
\docType{data}
\alias{cgg2022}
\title{Dataset of Chatziantoniou, Gabauer and Gupta (2022)}
\description{
For detailed information see: Chatziantoniou, I., Gabauer, D., & Gupta, R. (2021). Integration and Risk Transmission in the Market for Crude Oil: A Time-Varying Parameter Frequency Connectedness Approach.
}
\usage{data(cgg2022)}
\format{zoo data.frame}
\keyword{datasets}
