\name{cgp2024}
\docType{data}
\alias{cgp2024}
\title{Dataset of Cocca, Gabauer, and Pomberger (2024)}
\description{
For detailed information see: Cocca, T., Gabauer, D., & Pomberger, S. (2024). Clean energy market connectedness and investment strategies: New evidence from DCC-GARCH R2 decomposed connectedness measures. Energy Economics.
}
\usage{data(cgp2024)}
\format{zoo data.frame}
\keyword{datasets}
