\name{cgs2021}
\docType{data}
\alias{cgs2021}
\title{Dataset of Chatziantoniou, Gabauer and Stenfors (2021)}
\description{
For detailed information see: Chatziantoniou, I., Gabauer, D., & Stenfors, A. (2021). Interest rate swaps and the transmission mechanism of monetary policy: A quantile connectedness approach. Economics Letters, 204, 109891.
}
\usage{data(cgs2021)}
\format{zoo data.frame}
\keyword{datasets}
