\name{cgs2022}
\docType{data}
\alias{cgs2022}
\title{Dataset of Chatziantoniou, Gabauer and Stenfors (2022)}
\description{
For detailed information see: Chatziantoniou, I., Gabauer, D., & Stenfors, A. Independent Policy, Dependent Out-comes: A Game of Cross-Country Dom-inoes across European Yield Curves.
}
\usage{data(cgs2022)}
\format{zoo data.frame}
\keyword{datasets}
