% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalCorrelations.R
\name{ConditionalCorrelation}
\alias{ConditionalCorrelation}
\title{Partial Conditional Correlations}
\usage{
ConditionalCorrelation(Q)
}
\arguments{
\item{Q}{Variance-covariance matrix of dimension}
}
\value{
Get partial conditional correlations
}
\description{
Compute partial conditional correlations
}
\examples{
data(dy2012)
fit = VAR(dy2012, configuration=list(nlag=1))
pcc = ConditionalCorrelation(fit$Q)
}
\author{
David Gabauer
}
