% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectednessPartial.R
\name{ConnectednessPartial}
\alias{ConnectednessPartial}
\title{Partial Connectedness Measures}
\usage{
ConnectednessPartial(dca, group = c(1, 2), corrected = TRUE)
}
\arguments{
\item{dca}{Dynamic connectedness object}

\item{group}{Vector of group indices}

\item{corrected}{Should corrected TCI or normal TCI be used}
}
\value{
Get connectedness measures
}
\description{
This function results in partial connectedness measures
}
\examples{
#Replication of Gabauer and Gupta (2018)
#data("gg2018")
#dca = ConnectednessApproach(gg2018, model="TVP-VAR",
#                             connectedness="Time",
#                            nlag=1, nfore=10, window.size=200,
#                             VAR_config=list(TVPVAR=list(kappa1=0.99, kappa2=0.99,
#                                             prior="BayesPrior")))
#cd = ConnectednessAggregate(dca, groups=list("US"=c(1,2,3,4), "JP"=c(5,6,7,8)))
#plot(cd$NPSO_group[1,2,],type="l", ylim=c(-3,6), las=1, xaxs="i", xlab="", ylab="")
#lines(cd$NPSO_group[2,1,], col=2)
#lines(cd$NPSO_group[2,1,]-cd$NPSO_group[1,2,], col=3)
#abline(h=0, lty=3)
}
\references{
Gabauer, D., & Gupta, R. (2018). On the transmission mechanism of country-specific and international economic uncertainty spillovers: Evidence from a TVP-VAR connectedness decomposition approach. Economics Letters, 171, 63-71.
}
\author{
David Gabauer
}
