\name{dy2009}
\docType{data}
\alias{dy2009}
\title{Dataset of Diebold and Yilmaz (2009)}
\description{
For detailed information see: Diebold, F. X., & Yilmaz, K. (2009). Measuring financial asset return and volatility spillovers, with application to global equity markets. The Economic Journal, 119(534), 158-171.
}
\usage{data(dy2009)}
\format{A zoo data.frame containing 30x1141 observations.}
\source{Yahoo Finance}
\keyword{datasets}
