\name{dy2012}
\docType{data}
\alias{dy2012}
\title{Dataset of Diebold and Yilmaz (2012)}
\description{
For detailed information see: Diebold, F. X., & Yilmaz, K. (2012). Better to give than to receive: Predictive directional measurement of volatility spillovers. International Journal of forecasting, 28(1), 57-66.
}
\usage{data(dy2012)}
\format{A zoo data.frame containing 30x1141 observations.}
\source{Yahoo Finance}
\keyword{datasets}
