% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectednessTime.R
\name{DynamicConnectedness}
\alias{DynamicConnectedness}
\title{Diebold and Yilmaz (2009, 2012) connectedness approach}
\usage{
DynamicConnectedness(Phi, Sigma, nfore = 10, generalized = TRUE)
}
\arguments{
\item{Phi}{VAR coefficient matrix}

\item{Sigma}{Residual variance-covariance matrix}

\item{nfore}{H-step ahead forecast horizon}

\item{generalized}{Orthorgonalized/generalized FEVD}
}
\value{
Get connectedness measures
}
\description{
This function allows to calculate the Diebold and Yilmaz (2009, 2012) connectedness measures.
}
\examples{
#Replication of DY2009
#data("dy2009")
#fit = VAR(dy2009, configuration=list(nlag=2))
#dca = DynamicConnectedness(Phi=fit$B, Sigma=fit$Q, nfore=10, generalized=FALSE)
#dca$TABLE

#Replication of DY2012
#data("dy2012")
#fit = VAR(dy2012, configuration=list(nlag=4))
#dca = DynamicConnectedness(Phi=fit$B, Sigma=fit$Q, nfore=10, generalized=TRUE)
#dca$TABLE

#data("acg2020")
#prior = MinnesotaPrior(0.5, k=ncol(acg2020), nlag=1)
#fit = TVPVAR(acg2020, configuration=list(l=c(0.99,0.99), nlag=1, prior))
#dca = DynamicConnectedness(Phi=fit$B_t, Sigma=fit$Q_t, nfore=10, generalized=TRUE)
#dca$TABLE
}
\references{
Diebold, F. X., & Yilmaz, K. (2009). Measuring financial asset return and volatility spillovers, with application to global equity markets. The Economic Journal, 119(534), 158-171.\\
Diebold, F. X., & Yilmaz, K. (2012). Better to give than to receive: Predictive directional measurement of volatility spillovers. International Journal of Forecasting, 28(1), 57-66.
}
\author{
David Gabauer
}
