\name{g2020}
\docType{data}
\alias{g2020}
\title{Dataset of Gabauer (2020)}
\description{
For detailed information see: Gabauer, D. (2020). Volatility impulse response analysis for DCC-GARCH models: The role of volatility transmission mechanisms. Journal of Forecasting, 39(5), 788-796.
}
\usage{data(g2020)}
\format{zoo data.frame}
\keyword{datasets}
