% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GARCHtests.R
\name{GARCHtests}
\alias{GARCHtests}
\title{Univariate GARCH test statistics}
\usage{
GARCHtests(fit, lag = 20, prob = 0.05, conf.level = 0.9)
}
\arguments{
\item{fit}{Fitted univariate GARCH}

\item{lag}{Lag length of weighted Portmanteau statistics}

\item{prob}{The quantile (coverage) used for the VaR.}

\item{conf.level}{Confidence level of VaR test statistics}
}
\value{
Get best univariate GARCH
}
\description{
This function provides the results of multiple univariate GARCH test statistics
}
\references{
Ghalanos, A. (2014). rugarch: Univariate GARCH models, R package version 1.3-3.

Antonakakis, N., Chatziantoniou, I., & Gabauer, D. (2021). The impact of Euro through time: Exchange rate dynamics under different regimes. International Journal of Finance & Economics, 26(1), 1375-1408.
}
\author{
David Gabauer
}
