\name{gcat2022}
\docType{data}
\alias{gcat2022}
\title{Dataset of Chatziantoniou, Abakah, Gabauer & Tiwari (2022)}
\description{
For detailed information see: Chatziantoniou, I., Abakah, E. J., Gabauer, D., & Tiwari, A. K. (2022). Quantile time-frequency price connectedness between green bond, green equity, sustainable investments and clean energy markets: Implications for eco-friendly investors. Available at SSRN 3970746.
}
\usage{data(gcat2022)}
\format{zoo data.frame}
\keyword{datasets}
