\name{gg2018}
\docType{data}
\alias{gg2018}
\title{Dataset of Gabauer and Gupta (2018)}
\description{
For detailed information see, Gabauer, D., & Gupta, R. (2018). On the transmission mechanism of country-specific and international economic uncertainty spillovers: Evidence from a TVP-VAR connectedness decomposition approach. Economics Letters, 171, 63-71.
}
\usage{data(gg2018)}
\format{zoo data.frame}
\keyword{datasets}
