\name{gghm2022}
\docType{data}
\alias{gghm2022}
\title{Dataset of Gabauer, Gupta, Haradik and Miller (2020)}
\description{
For detailed information see: Gabauer, D., Gupta, R., Marfatia, H., and Miller, S. M. (2020). Estimating us housing price network connectedness: Evidence from dynamic elastic net, lasso, and ridge vector autoregressive models.
}
\usage{data(gghm2022)}
\format{zoo data.frame}
\keyword{datasets}
