% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GiniConnectedness.R
\name{GiniConnectedness}
\alias{GiniConnectedness}
\title{Gini-Based Connectedness Measures}
\usage{
GiniConnectedness(dca)
}
\arguments{
\item{dca}{Dynamic connectedness object}
}
\value{
Get Gini-based connectedness measures
}
\description{
This function results in Gini-based connectedness measures
}
\examples{
\donttest{
data("dy2012")
dca = ConnectednessApproach(dy2012,
                            nlag=1,
                            nfore=20,
                            model="VAR",
                            connectedness="Time",
                            corrected=TRUE)
gca = GiniConnectedness(dca)
}
}
\references{
Gabauer, D., Stenfors, A., & Vinco, V. (2026). Eurozone sectoral inflation networks. International Review of Economics and Finance, 109, 105405.
}
\author{
David Gabauer
}
