\name{jcggh2022}
\docType{data}
\alias{jcggh2022}
\title{Dataset of Juncal, Chatziantoniou, Gabauer, Garcia & Hardik (2022)}
\description{
For detailed information see: Juncal, C., Chatziantoniou, I., Gabauer, D., De Gracia, F. P., & Hardik, M. (2022). Dynamic spillovers across precious metals and energy realized volatilities: Evidence from quantile extended joint connectedness measures.}
\usage{data(jcggh2022)}
\format{zoo data.frame}
\keyword{datasets}
