\name{lw2021}
\docType{data}
\alias{lw2021}
\title{Dataset of Lastrapes and Wiesen (2021)}
\description{
For detailed information see: Lastrapes, W. D., & Wiesen, T. F. (2021). The joint spillover index. Economic Modelling, 94, 681-691.
}
\usage{data(lw2021)}
\format{zoo data.frame}
\keyword{datasets}
