% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinnesotaPrior.R
\name{MinnesotaPrior}
\alias{MinnesotaPrior}
\title{Minnesota Prior}
\usage{
MinnesotaPrior(gamma = 0.1, k, nlag)
}
\arguments{
\item{gamma}{Diagonal value of variance-covariance matrix}

\item{k}{Number of series}

\item{nlag}{Lag length}
}
\value{
Get Minnesota Prior
}
\description{
Get Minnesota Prior
}
\examples{
\donttest{
prior = MinnesotaPrior(0.1, k=4, nlag=1)
}
}
\references{
Koop, G., & Korobilis, D. (2010). Bayesian multivariate time series methods for empirical macroeconomics. Now Publishers Inc.
}
\author{
David Gabauer
}
