% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_influence.R
\name{plot_influence}
\alias{plot_influence}
\title{Dynamic influence connectedness plot}
\usage{
plot_influence(
  ca,
  col = "steelblue4",
  save = FALSE,
  lower = NULL,
  upper = NULL,
  ...
)
}
\arguments{
\item{ca}{Connectedness object}

\item{col}{Color}

\item{save}{Save plot as pdf}

\item{lower}{Lower limit}

\item{upper}{Upper limit}

\item{...}{Arguments to be passed to methods, such as graphical parameters (see par).}
}
\value{
Return connectedness plot
}
\description{
Visualize dynamic influence connectedness
}
