% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_npdc.R
\name{plot_npdc}
\alias{plot_npdc}
\title{Dynamic net pairwise transmission plot}
\usage{
plot_npdc(ca, save = FALSE, path = "./Results", ...)
}
\arguments{
\item{ca}{Connectedness object}

\item{save}{Save plot as pdf}

\item{path}{Path where plots should be saved}

\item{...}{Arguments to be passed to methods, such as graphical parameters (see par).}
}
\value{
Return connectedness plot
}
\description{
Visualize dynamic net total directional connectedness
}
