% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tci.R
\name{plot_tci}
\alias{plot_tci}
\title{Dynamic total connectedness plot}
\usage{
plot_tci(
  ca,
  save = FALSE,
  path = "./Results",
  ylim = c(NULL, NULL),
  corrected = TRUE,
  ...
)
}
\arguments{
\item{ca}{Connectedness object}

\item{save}{Save plot as pdf}

\item{path}{Path where plots should be saved}

\item{ylim}{A vector including the lower and upper limit of the y-axis}

\item{corrected}{Visualizing standard or corrected TCI values}

\item{...}{Arguments to be passed to methods, such as graphical parameters (see par).}
}
\value{
Return connectedness plot
}
\description{
Visualize dynamic total connectedness
}
