% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNetwork.R
\name{PlotNetwork}
\alias{PlotNetwork}
\title{Network plot}
\usage{
PlotNetwork(dca, method = "NPDC", path = NULL, name_length = NULL, ...)
}
\arguments{
\item{method}{Either visualizing NPDC or PCI}

\item{path}{Path where plots should be saved}

\item{name_length}{Length of variable names in the network plot}

\item{...}{Arguments to be passed to methods, such as graphical parameters (see par).}

\item{x}{NPDC or PCI matrix}
}
\value{
Return connectedness plot
}
\description{
Visualize net pairwise or pairwise connectedness measures
}
