% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPCI.R
\name{PlotPCI}
\alias{PlotPCI}
\title{Dynamic pairwise connectedness plot}
\usage{
PlotPCI(
  dca,
  ca = NULL,
  path = NULL,
  ylim = c(NULL, NULL),
  selection = NULL,
  ...
)
}
\arguments{
\item{dca}{Connectedness object}

\item{ca}{Compare dca object with a single connectedness object or a list of of connectedness objects}

\item{path}{Path where plots should be saved}

\item{ylim}{A vector including the lower and upper limit of the y-axis}

\item{selection}{Indidcator of the illustrated series}

\item{...}{Arguments to be passed to methods, such as graphidcal parameters (see par).}
}
\value{
Return connectedness plot
}
\description{
Visualize dynamic pairwise connectedness
}
