% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileCorrelation.R
\name{QuantileCorrelation}
\alias{QuantileCorrelation}
\title{Quantile correlation}
\usage{
QuantileCorrelation(x, tau = 0.5, method = "fn")
}
\arguments{
\item{x}{zoo data matrix}

\item{tau}{quantile between 0 and 1}

\item{method}{Either "br", "fn" or "sfn". Default is "fn"}
}
\value{
Get quantile correlations
}
\description{
This function computes the quantile correlation coefficient proposed by Choi and Shin (2022).
}
\examples{
\donttest{
data("g2020")
fit = QuantileCorrelation(g2020, tau=0.5)
}
}
\references{
Choi, J. E., & Shin, D. W. (2022). Quantile correlation coefficient: A new tail dependence measure. Statistical Papers, 63(4), 1075-1104.
}
\author{
David Gabauer
}
