% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2Correlations.R
\name{R2Correlations}
\alias{R2Correlations}
\title{R2 decomposed connectedness from correlations}
\usage{
R2Correlations(R)
}
\arguments{
\item{R}{zoo correlation data matrix}
}
\value{
Get R2 connectedness measures from correlation matrix
}
\description{
This function computes the R2 decomposed connectedness measures from correlations
}
\references{
Naeem, M. A., Chatziantoniou, I., Gabauer, D., & Karim, S. (2023). Measuring the G20 Stock Market Return Transmission Mechanism: Evidence From the R2 Connectedness Approach. International Review of Financial Analysis.

Balli, F., Balli, H. O., Dang, T. H. N., & Gabauer, D. (2023). Contemporaneous and lagged R2 decomposed connectedness approach: New evidence from the energy futures market. Finance Research Letters, 57, 104168.
}
\author{
David Gabauer
}
