% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2XConnectedness.R
\name{R2XConnectedness}
\alias{R2XConnectedness}
\title{R2 connectedness approach with exogenous input}
\usage{
R2XConnectedness(
  x,
  z = NULL,
  window.size = NULL,
  nlag = 1,
  xlag = 0,
  method = "pearson",
  relative = FALSE
)
}
\arguments{
\item{x}{zoo data matrix}

\item{z}{zoo data matrix for exogenous input}

\item{window.size}{Rolling-window size or Bayes Prior sample size}

\item{nlag}{Lag length}

\item{xlag}{Lag length of exogenous variables}

\item{method}{Methods for R2 connectedness are:"pearson", "spearman", or "kendall". "pearson" is default. Methods for pseudo R2 quantile connectedness are: "lasso", "br", "fn", or "sfn". "lasso" is default.}

\item{relative}{Boolean whether relative or absolute R2 should be used}
}
\value{
Get R2 connectedness measures
}
\description{
This function computes the R2 connectedness measures with exogenous input
}
\examples{
\donttest{
data("dy2012")
dca = R2XConnectedness(x=dy2012[,1:2, drop=F], z=dy2012[,3:4, drop=F], window.size=NULL, nlag=1, xlag=1, method="pearson")
dca$TABLE
}
}
\references{
Naeem, M. A., Chatziantoniou, I., Gabauer, D., & Karim, S. (2023). Measuring the G20 Stock Market Return Transmission Mechanism: Evidence From the R2 Connectedness Approach. International Review of Financial Analysis.

Balli, F., Balli, H. O., Dang, T. H. N., & Gabauer, D. (2023). Contemporaneous and lagged R2 decomposed connectedness approach: New evidence from the energy futures market. Finance Research Letters, 57, 104168.

Ferrer, R., Shahzad, S. J. H., Furió, D., & Benammar, R. (2025). Systemic Risk in the Tails: Contemporaneous Transmission and Spillover Dynamics in European Renewable Energy Equities.
}
\author{
David Gabauer
}
