% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustCovariance.R
\name{RobustCovariance}
\alias{RobustCovariance}
\title{Robust Covariance}
\usage{
RobustCovariance(x, method = "pearson")
}
\arguments{
\item{x}{zoo data matrix}

\item{method}{Either "mcd", "mve", "ogk", "tyler", "sccm" (covariance), "pearson", "spearman" and "kendall" (correlation) or a combination of the two.}
}
\value{
Estimate robust covariance matrix
}
\description{
Estimation of robust covariance
}
\examples{
data("dy2012")
RobustCovariance(dy2012, method=c("pearson"))
RobustCovariance(dy2012, method=c("mcd", "pearson"))
}
\author{
David Gabauer
}
