% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStatistics.R
\name{SummaryStatistics}
\alias{SummaryStatistics}
\title{Summary Statistics}
\usage{
SummaryStatistics(
  x,
  portmanteau = c("Ljung-Box", "Box-Pierce", "Monti"),
  correlation = c("kendall", "spearman", "pearson"),
  nlag = 20,
  digit = 3
)
}
\arguments{
\item{x}{zoo data matrix}

\item{portmanteau}{portmanteau statistics: "Box-Pierce", "Ljung-Box", "Monti"}

\item{correlation}{correlation coefficient: "pearson", "kendall", "spearman".}

\item{nlag}{number of lags for Weighted Portmanteau statistics}

\item{digit}{digit Number of decimal places}
}
\value{
Get summary statistics
}
\description{
Get comprehensive summary statistics
}
\examples{
\donttest{
data(dy2012)
SummaryStatistics(dy2012)
}
}
\author{
David Gabauer
}
