% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UninformativePrior.R
\name{UninformativePrior}
\alias{UninformativePrior}
\title{Uninformative Prior}
\usage{
UninformativePrior(k, nlag)
}
\arguments{
\item{k}{Number of series}

\item{nlag}{Lag length}
}
\value{
Get Uninformative Prior
}
\description{
Get Uninformative Prior
}
\examples{
prior = UninformativePrior(k=4, nlag=1)
}
\references{
Koop, G., & Korobilis, D. (2010). Bayesian multivariate time series methods for empirical macroeconomics. Now Publishers Inc.
}
\author{
David Gabauer
}
