% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wold.R
\name{Wold}
\alias{Wold}
\title{Wold representation theorem}
\usage{
Wold(x, nfore = 10)
}
\arguments{
\item{x}{VAR coefficients}

\item{nfore}{H-step ahead forecast horizon}
}
\value{
Get VMA coefficients
}
\description{
Transform VAR to VMA coefficients
}
\examples{
data(dy2012)
fit = VAR(dy2012, configuration=list(nlag=1))
wold = Wold(fit$B, nfore=10)
}
\references{
Sims, C. A. (1980). Macroeconomics and reality. Econometrica, 1-48.
}
\author{
David Gabauer
}
