% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateDCCGARCH.R
\name{BivariateDCCGARCH}
\alias{BivariateDCCGARCH}
\title{Bivariate DCC-GARCH}
\usage{
BivariateDCCGARCH(
  data,
  spec,
  copula = c("mvnorm", "mvt"),
  method = c("Kendall", "ML"),
  transformation = c("parametric", "empirical", "spd"),
  time.varying = TRUE,
  asymmetric = FALSE
)
}
\arguments{
\item{data}{zoo dataset}

\item{spec}{A cGARCHspec A cGARCHspec object created by calling cgarchspec.}

\item{copula}{"mvnorm" or "mvt" (see, rmgarch package)}

\item{method}{"Kendall" or "ML" (see, rmgarch package)}

\item{transformation}{"parametric", "empirical" or "spd" (see, rmgarch package)}

\item{time.varying}{Boolean value to either choose DCC-GARCH or CCC-GARCH}

\item{asymmetric}{Whether to include an asymmetry term to the DCC model (thus estimating the aDCC).}
}
\value{
Estimate Bivariate DCC-GARCH
}
\description{
This function multiple Bivariate DCC-GARCH models that captures more accurately conditional covariances and correlations
}
\examples{
#data("g2020")
#uspec = ugarchspec()
#fit = BivariateDCCGARCH(data=g2020, spec=replicate(nol(g2020), uspec))
}
\references{
Antonakakis, N., Chatziantoniou, I. & Gabauer, D. (2022).
}
\author{
David Gabauer
}
