\name{cegg2022}
\docType{data}
\alias{cegg2022}
\title{Dataset of Chatziantoniou, Elsayed, Gabauer and Gozgor (2022)}
\description{
For detailed information see: Chatziantoniou, I., Elsayed, AH., Gabauer, D. and Gozgor, G. (2021). Oil price shocks and exchange rate dynamics: New evidence from internal, external and partial connectedness measures for oil importing and exporting countries
}
\usage{data(cegg2022)}
\format{zoo data.frame}
\keyword{datasets}
