% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InclusiveConnectedness.R
\name{InclusiveConnectedness}
\alias{InclusiveConnectedness}
\title{Inclusive Connectedness Measures}
\usage{
InclusiveConnectedness(dca, group = c(1, 2), start = NULL, end = NULL)
}
\arguments{
\item{dca}{Dynamic connectedness object}

\item{group}{Vector of group indices}

\item{start}{Start index}

\item{end}{End index}
}
\value{
Get connectedness measures
}
\description{
This function results in inclusive connectedness measures
}
\examples{
\donttest{
data("cegg2022")
dca = ConnectednessApproach(cegg2022,
                            model="VAR",
                            connectedness="Time",
                            nlag=1,
                            nfore=20,
                            corrected=TRUE)
inc = InclusiveConnectedness(dca, group=c(1,2,3))
}
}
\references{
Chatziantoniou, I., Elsayed, A. H., Gabauer, D., & Gozgor, G. (2023). Oil price shocks and exchange rate dynamics: Evidence from decomposed and partial connectedness measures for oil importing and exporting economies. Energy Economics, 120, 106627.
}
\author{
David Gabauer
}
