% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateHedgingPortfolio.R
\name{MultivariateHedgingPortfolio}
\alias{MultivariateHedgingPortfolio}
\title{Multivariate Hedging Portfolio}
\usage{
MultivariateHedgingPortfolio(
  x,
  H,
  method = c("cumsum", "cumprod"),
  statistics = c("Fisher", "Bartlett", "Fligner-Killeen", "Levene", "Brown-Forsythe"),
  metric = "StdDev",
  digit = 2
)
}
\arguments{
\item{x}{zoo return matrix (in percentage)}

\item{H}{Residual variance-covariance, correlation or pairwise connectedness matrix}

\item{method}{Cumulative sum or cumulative product}

\item{statistics}{Hedging effectiveness statistic}

\item{metric}{Risk measure of Sharpe Ratio (StdDev, VaR, or CVaR)}

\item{digit}{Number of decimal places}
}
\value{
Get hedge ratios
}
\description{
This function calculates the hedge ratios of Kroner and Sultan (1993)
}
\examples{
data("g2020")
fit = VAR(g2020, configuration=list(nlag=1))
hr = MultivariateHedgingPortfolio(g2020/100, fit$Q)
hr$TABLE
}
\author{
David Gabauer
}
