% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_net.R
\name{plot_net}
\alias{plot_net}
\title{Dynamic net total directional connectedness plot}
\usage{
plot_net(ca, save = FALSE, path = "./Results", ylim = c(NULL, NULL), ...)
}
\arguments{
\item{ca}{Connectedness object}

\item{save}{Save plot as pdf}

\item{path}{Path where plots should be saved}

\item{ylim}{A vector including the lower and upper limit of the y-axis}

\item{...}{Arguments to be passed to methods, such as graphical parameters (see par).}
}
\value{
Return connectedness plot
}
\description{
Visualize dynamic net total directional connectedness
}
