% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_npdc.R
\name{plot_npdc}
\alias{plot_npdc}
\title{Dynamic net pairwise transmission plot}
\usage{
plot_npdc(ca, col = "steelblue4", save = FALSE, ...)
}
\arguments{
\item{ca}{Connectedness object}

\item{col}{Color}

\item{save}{Save plot as pdf}

\item{...}{Arguments to be passed to methods, such as graphical parameters (see par).}
}
\value{
Return connectedness plot
}
\description{
Visualize dynamic net total directional connectedness
}
