% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tci.R
\name{plot_tci}
\alias{plot_tci}
\title{Dynamic total connectedness plot}
\usage{
plot_tci(ca, col = "steelblue4", save = FALSE, lower = NULL, upper = NULL, ...)
}
\arguments{
\item{ca}{Connectedness object}

\item{col}{Color}

\item{save}{Save plot as pdf}

\item{lower}{Lower limit}

\item{upper}{Upper limit}

\item{...}{Arguments to be passed to methods, such as graphical parameters (see par).}
}
\value{
Return connectedness plot
}
\description{
Visualize dynamic total connectedness
}
