% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNET.R
\name{PlotNET}
\alias{PlotNET}
\title{Dynamic net total directional connectedness plot}
\usage{
PlotNET(dca, ca = NULL, path = NULL, ylim = c(NULL, NULL), ...)
}
\arguments{
\item{dca}{Connectedness object}

\item{ca}{Compare dca object with a single connectedness object or a list of of connectedness objects}

\item{path}{Path where plots should be saved}

\item{ylim}{A vector including the lower and upper limit of the y-axis}

\item{...}{Arguments to be passed to methods, such as graphical parameters (see par).}
}
\value{
Return connectedness plot
}
\description{
Visualize dynamic net total directional connectedness
}
