% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNetwork.R
\name{PlotNetwork}
\alias{PlotNetwork}
\title{Network plot}
\usage{
PlotNetwork(
  dca,
  method = "NPDC",
  path = NULL,
  name_length = NULL,
  threshold = 0,
  width = 10,
  height = 10,
  ...
)
}
\arguments{
\item{dca}{Connectedness object}

\item{method}{Either visualizing NPDC or PCI}

\item{path}{Path where plots should be saved}

\item{name_length}{Length of variable names in the network plot}

\item{threshold}{Threshold for bivariate connections between 0 and 1}

\item{width}{The width of the graphics region in inches}

\item{height}{The height of the graphics region in inches}

\item{...}{Arguments to be passed to methods, such as graphical parameters (see par).}
}
\value{
Return connectedness plot
}
\description{
Visualize net pairwise or pairwise connectedness measures
}
