% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotNPT.R
\name{PlotNPT}
\alias{PlotNPT}
\title{Dynamic net pairwise transmission plot}
\usage{
PlotNPT(dca, ca = NULL, path = NULL, ...)
}
\arguments{
\item{dca}{Connectedness object}

\item{ca}{Compare dca object with a single connectedness object or a list of of connectedness objects}

\item{path}{Path where plots should be saved}

\item{...}{Arguments to be passed to methods, such as graphidcal parameters (see par).}
}
\value{
Return connectedness plot
}
\description{
Visualize dynamic net total directional connectedness
}
