% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2Connectedness.R
\name{R2Connectedness}
\alias{R2Connectedness}
\title{R2 connectedness approach}
\usage{
R2Connectedness(
  x,
  window.size = NULL,
  nlag = 1,
  tau = NULL,
  method = "pearson",
  relative = FALSE,
  lambda = 0
)
}
\arguments{
\item{x}{zoo data matrix}

\item{window.size}{Rolling-window size or Bayes Prior sample size}

\item{nlag}{Lag length}

\item{tau}{Quantile between 0 and 1}

\item{method}{Methods for R2 connectedness are:"pearson", "spearman", or "kendall". "pearson" is default. Methods for pseudo R2 quantile connectedness are:"br", "fn", or "sfn". "fn" is default.}

\item{relative}{Boolean whether relative or absolute R2 should be used}

\item{lambda}{The correlation shrinkage intensity (range 0-1). If lambda is not specified (the default) it is estimated using an analytic formula from Sch\"afer and Strimmer (2005) - see details below. For lambda=0 the empirical correlations are recovered.}
}
\value{
Get R2 connectedness measures
}
\description{
This function computes the R2 connectedness measures as well as peudo R2 quantile connectedness measures
}
\examples{
\donttest{
data("dy2012")
dca = R2Connectedness(dy2012, window.size=NULL, nlag=0, method="pearson")
dca$TABLE
}
}
\references{
Naeem, M. A., Chatziantoniou, I., Gabauer, D., & Karim, S. (2023). Measuring the G20 Stock Market Return Transmission Mechanism: Evidence From the R2 Connectedness Approach. International Review of Financial Analysis.

Balli, F., Balli, H. O., Dang, T. H. N., & Gabauer, D. (2023). Contemporaneous and lagged R2 decomposed connectedness approach: New evidence from the energy futures market. Finance Research Letters, 57, 104168.

Ferrer, R., Shahzad, S. J. H., Furió, D., & Benammar, R. (2025). Systemic Risk in the Tails: Contemporaneous Transmission and Spillover Dynamics in European Renewable Energy Equities.
}
\author{
David Gabauer
}
